Refactored exception handling all over the library
