+++ src/main/java/com/spotify/docker/client/LoggingPullHandler.java
@@ -17,6 +17,9 @@
 
 package com.spotify.docker.client;
 
+import com.spotify.docker.client.exceptions.DockerException;
+import com.spotify.docker.client.exceptions.ImageNotFoundException;
+import com.spotify.docker.client.exceptions.ImagePullFailedException;
 import com.spotify.docker.client.messages.ProgressMessage;
 
 import org.slf4j.Logger;
