+++ src/main/java/com/spotify/docker/client/exceptions/ExecNotFoundException.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2014 Oleg Poleshuk.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.spotify.docker.client.exceptions;
+
+public class ExecNotFoundException extends NotFoundException {
+
+  private final String execId;
+
+  public ExecNotFoundException(final String execId, final Throwable cause) {
+    super("Exec not found: " + execId, cause);
+    this.execId = execId;
+  }
+
+  public ExecNotFoundException(final String execId) {
+    this(execId, null);
+  }
+
+  public String getExecId() {
+    return execId;
+  }
+}
