+++ src/main/java/com/spotify/docker/client/exceptions/ImageNotFoundException.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2014 Spotify AB.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.spotify.docker.client.exceptions;
+
+public class ImageNotFoundException extends NotFoundException {
+
+  private final String image;
+
+  public ImageNotFoundException(final String image, final Throwable cause) {
+    super("Image not found: " + image, cause);
+    this.image = image;
+  }
+
+  public ImageNotFoundException(final String image, final String message) {
+    super("Image not found: " + image + ": " + message);
+    this.image = image;
+  }
+
+  public ImageNotFoundException(final String image) {
+    super("Image not found: " + image);
+    this.image = image;
+  }
+
+  public String getImage() {
+    return image;
+  }
+}
